(original rev 2160) merge trunk to staging rev 2156
git-svn-id: file:///tmp/current/omni/trunk@1920 2ffc8166-1744-45da-a82a-d6ea10f7896b

+++ omni/src/utility/stringHelpers.h
@@ -10,7 +10,7 @@ class StringHelpers
 	static QString getStringFromSegmentSet( const OmSegIDsSet & data_set );
 	static QString getStringFromIDset( const OmIDsSet & data_set );
 	static QString getStringFromStringList( const QStringList & data_set );
-	static QString commaDeliminateNumber( const int num );
+	static QString commaDeliminateNumber( const long long num );
 	static unsigned int getUInt( QString arg );
 	static double getDouble( QString arg );
 };
